requirements.txt
absl-py==0.7.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.3.2
defer==1.0.6
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
Django==2.0.13
django-ipware==2.1.0
django-leaflet==0.19.0
django-qsessions==0.2.1
frozendict==1.2
gast==0.2.2
gpg==1.10.0
grpcio==1.18.0
h5py==2.9.0
httplib2==0.9.2
idna==2.6
imutils==0.5.2
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
netifaces==0.10.4
numpy==1.16.1
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.0.25
parso==0.3.3
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
pipenv==2018.11.26
prompt-toolkit==2.0.8
protobuf==3.6.1
psycopg2==2.8.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
pygobject==3.26.1
pyknow==1.7.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-firebase-gae==1.0.1
python-magic==0.4.16
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
schema==0.6.7
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
tornado==5.1.1
traitlets==4.3.2
ua-parser==0.8.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.22
usb-creator==0.3.3
user-agents==2.0
vboxapi==1.0
virtualenv==16.5.0
virtualenv-clone==0.5.3
wadllib==1.3.2
wcwidth==0.1.7
WebOb==1.6.1
Werkzeug==0.14.1
xkit==0.0.0
zope.interface==4.3.2
